==> 01-Data-Import.ipynb <==
# Data cleaning and pre-processing

Project flow:  
1. Topic modeling to develop the aspect categories  
2. Reshape long: Review id with multiple aspects per review  
3. Finding the aspect words in sentences and classify to aspect category
**HOW**  
  
4. Computing the polarity (sentiment) for each aspect in each entry  
5. Create a sentence for each aspect/sentiment
6. Create a summary of each review - sentence order weighted by aspect topic weight  

**IMPORT SUPPORTING PACKAGES**

#!python -m spacy download en
import spacy
nlp = spacy.load('en')
import nltk
from nltk.corpus import stopwords
import json
import boto3
from pandas.io.json import json_normalize
import pandas as pd
import numpy as np
import pickle
import nltk
import re
import string
from nltk.corpus import stopwords

nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
lemmatizer = nltk.stem.WordNetLemmatizer()
w_tokenizer = nltk.tokenize.WhitespaceTokenizer()

**DEFINE HELPER FUNCTIONS**

def read_json_file_from_S3(filename):
    s3 = boto3.resource('s3',
                        aws_access_key_id='AKIAJLCANPRLUUJ3MR5A', #provide your aws_access_key_id
                        aws_secret_access_key= 'LgvENqvcUu6HDorEGBZqj6sZv6Soiy3H9zirUo6k') #provide your aws_secret_access_key

    content_object = s3.Object('mertisyelpdataset', filename)
    rawdata = []
    for line in content_object.get()['Body'].read().splitlines():
        rawdata.append(line)
    jsondata = list(map(lambda x: json.loads(x.decode('utf-8')), rawdata))
    return json_normalize(jsondata)

def split_sentences(review_df):
    a=review_df[['text','review_id']]
    b = pd.DataFrame(a.text.str.split('.').tolist(), index=a.index).stack()
    b = pd.DataFrame(b).reset_index()
    b.columns = ['review_index', 'sentence_index', 'sentence']
    b = b[b['sentence'].isna() == False]
    b = b[b['sentence'] != ''].reset_index(drop = True)
    sentences = pd.merge(b, a, how='left', left_on='review_index', right_index=True)
    sentences = sentences.drop(['text'], axis=1)
    return(sentences)

nlp = spacy.load('en', disable=['parser', 'ner'])
def clean_sentences(sentences):
    sentences = [re.sub("\n|\r", "", s) for s in sentences]
    
    sentences = [s.lower() for s in sentences] # lower case    
    sentences = [s.translate(str.maketrans('', '', string.punctuation)) for s in sentences] # remove punctuation
    
    allowed_postags=['NOUN']
    def lemmatize_alphanumeric(sentence):
        words = [lemmatizer.lemmatize(word) for word in sentence.split() if word.isalpha()
                and nlp.vocab[word].is_stop == False]
        sentence = ' '.join(words)
        return(sentence)
    
    sentences = [lemmatize_alphanumeric(s) for s in sentences] # lemmatize and remove non alpha-numeric characters
    sentences = [w_tokenizer.tokenize(s) for s in sentences]
    return(sentences)

def get_nouns(text):
    allowed_postags=['NOUN']
    doc = nlp(' '.join(text))
    nouns = [word for word in doc if word.pos_ in allowed_postags]
    return(nouns)

#review_df = read_json_file_from_S3('review.json')
#pd.to_pickle(review_df, 'reviews.pkl')

#bus_df = read_json_file_from_S3('business.json')
#pd.to_pickle(bus_df, 'business.pkl')

**1. Business df for pizza categories only**

bus_df = pd.read_pickle('business.pkl')


bus_df.columns

bus_df['categories'] = [str(X).lower() for X in bus_df['categories']]
cats = ['pizza, restaurants','restaurants, pizza','italian, restaurants, pizza',
        'restaurants, italian, pizza','restaurants, pizza, italian','pizza, restaurants, italian',
        'pizza, italian, restaurants','italian, pizza, restaurants']
bus_za = bus_df[bus_df.categories.isin(cats)]


bus_za.head(2)

bus_za.shape

del(bus_df)

**Dataframe of all pizza bussiness info**

bus_id_all = bus_za['business_id'].tolist()

len(bus_id_all)

pd.to_pickle(bus_za, 'pizza_bussinesses_df.pkl')
pd.to_pickle(bus_id_all, 'pizza_businesses_ids.pkl')

**Dataframe of business sample**

sample = pd.DataFrame(bus_za.business_id.unique()).sample(250, random_state=80)
sample = sample[0].tolist()
sample[0:5]

bus_za_sample = bus_za[bus_za['business_id'].isin(sample)]

len(bus_za_sample)

pd.to_pickle(bus_za_sample, 'pizza_businesses_df_sample.pkl')
pd.to_pickle(sample, 'pizza_businesses_ids_sample.pkl')

**2. REVIEWS**

review_df = pd.read_pickle('reviews.pkl')

review_df.head(3)

review_df = review_df[review_df['business_id'].isin(bus_id_all)]

review_df['cleaned'] = clean_sentences(review_df.text)
review_df['cleaned'] = [get_nouns(s) for s in review_df['cleaned']]
review_df['cleaned'] = [w_tokenizer.tokenize(str(X).strip('][').replace(',','')) for X in review_df['cleaned']]

review_sample = review_df[review_df['business_id'].isin(sample)]

review_sample_ids = list(review_sample['review_id'])

pd.to_pickle(review_df, 'pizza_reviews_df.pkl')
pd.to_pickle(review_sample, 'pizza_reviews_df_sample.pkl')

review_sample = pd.read_pickle('pizza_reviews_df_sample.pkl')

len(review_sample.business_id.unique())

print(review_sample['review_id'][review_sample['stars']==3].iloc[800])

**3. SENTENCES FROM REVIEWS**

review_df.head(2)

sentences = split_sentences(review_df)

review_df.shape

len(sentences.review_id.unique())

sentences['cleaned'] = clean_sentences(sentences.sentence)
sentences['cleaned'] = [get_nouns(s) for s in sentences['cleaned']]
#sentences['cleaned'] = [w_tokenizer.tokenize(X.strip('][').replace(',','')) for X in sentences['cleaned']]

sentences['cleaned'] = [str(X) for X in sentences['cleaned']]

sentences_sample = sentences[sentences['review_id'].isin(review_sample_ids)]

sentences_sample.head()

pd.to_pickle(sentences, 'pizza_sentences.pkl')
pd.to_pickle(sentences_sample, 'pizza_sentences_sample.pkl')

**END NOTEBOOK**

***
