==> 03-LDA-All-Sentences.ipynb <==
# Topic modeling 1

import spacy
nlp = spacy.load('en')
import nltk
from nltk.corpus import stopwords
import pandas as pd
import numpy as np
import pickle

import re

import nltk
import re
import string
from nltk.corpus import stopwords

nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
lemmatizer = nltk.stem.WordNetLemmatizer()
w_tokenizer = nltk.tokenize.WhitespaceTokenizer()

import gensim
import pyLDAvis
from pyLDAvis import gensim as gensimvis
import spacy

import logging
from tqdm import tqdm
from pprint import pprint

**IMPORT DATA**

df = pd.read_pickle('pizza_sentences.pkl')


#df['cleaned'] = [w_tokenizer.tokenize(X.strip('][').replace(',','')) for X in df['cleaned']]
df.head()

df.shape

**SENTIMENT ANALYSIS**

from pycorenlp import StanfordCoreNLP
#in terminal:
#cd stanford-corenlp-full-2018-10-05
#java -mx5g -cp "*" edu.stanford.nlp.pipeline.StanfordCoreNLPServer -timeout 10000

nlp = StanfordCoreNLP('http://localhost:9000')

res = nlp.annotate(list(df.sentence)[1111],
                   properties={'annotators': 'sentiment',
                               'outputFormat': 'json',
                               'timeout': 1000,
                   })
for s in res["sentences"]:
    print("%d: '%s': %s %s" % (
        s["index"],
        " ".join([t["word"] for t in s["tokens"]]),
        s["sentimentValue"], s["sentiment"]))

def get_sentiment(sentence):
    try:
        res = nlp.annotate(sentence,
                       properties={'annotators': 'sentiment',
                                   'outputFormat': 'json',
                                   'timeout': 1000,
                       })
        result = (res['sentences'][0]['sentimentValue'], res['sentences'][0]['sentiment'])
    except Exception as e:
        result = ('error', e)
    return(result)

df = df[df['sentence'].isna()==False]
df['sentiment'] = [get_sentiment(X) for X in df['sentence']]
df['sentiment_val'] = [(X[0]) for X in df.sentiment]
df['sentiment_val'] = [int(X) if X != 'error' else np.nan for X in df.sentiment_val]


pd.to_pickle(df,'pizza_sentences.pkl')

**TOPIC MODELING**

new_stopwords = ['way', 'honey', 'otto', 'pick', 'end', 'vega', 'libretto', 'hour',
                 'place', 'vitos', 'wish', 'point', 'thats', 'bit', 'piece', 'word',
                 'review', 'door', 'guy', 'ill', 'alla', 'im', 'yelp', 'type', 'montreal',
                 'phone', 'aria', 'day', 'phoenix', 'star', 'ive', 'lot', 'need', 'table',
                 'thing', 'mario', 'year', 'spot', 'rest']
nlp = spacy.load('en', disable=['parser', 'ner'])
for s in new_stopwords:
    nlp.Defaults.stop_words.add(s)
    lex = nlp.vocab[s]
    lex.is_stop = True
    nlp.vocab[s].is_stop = True


def clean_stopwords(tokenized_text):
    tokenized_text = [t for t in tokenized_text if nlp.vocab[t].is_stop == False]
    return(tokenized_text)

def get_topics_lda(tokenized_text, num_topics):
    id2word = gensim.corpora.Dictionary(tokenized_text)
    corpus = [id2word.doc2bow(t) for t in tokenized_text]
    lda_model = gensim.models.ldamulticore.LdaMulticore(corpus=corpus,
                                           id2word=id2word,
                                           num_topics=num_topics, 
                                           random_state=100,
                                           chunksize=100,
                                           passes=128,
                                           per_word_topics=True)
    return(lda_model)

df['cleaned'] = [clean_stopwords(t) for t in df['cleaned']]
id2word = gensim.corpora.Dictionary(df.cleaned)
corpus = [id2word.doc2bow(t) for t in df.cleaned]

lda_model_whole = get_topics_lda(df.cleaned, 5)
lda_model_whole.print_topics(num_words=20)

pd.to_pickle(lda_model_whole, 'lda_all.pkl')

lda_model_whole = pd.read_pickle('lda_all.pkl')

#get_document_topics = [lda_model_whole.get_document_topics(item) for item in corpus]

get_document_topics

lda_model_whole.print_topics(num_words=20)

**END NOTEBOOK**
