==> 05-Summaries.ipynb <==
# Generate summaries for each restaurant

import spacy
nlp = spacy.load('en')
import nltk
from nltk.corpus import stopwords
import pandas as pd
import numpy as np
import pickle

import re

import nltk
import re
import string
from nltk.corpus import stopwords

nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
lemmatizer = nltk.stem.WordNetLemmatizer()
w_tokenizer = nltk.tokenize.WhitespaceTokenizer()

from nltk.corpus import stopwords

import gensim
import pyLDAvis
from pyLDAvis import gensim as gensimvis
import spacy

from gensim.models.coherencemodel import CoherenceModel
from gensim.models.ldamodel import LdaModel
from gensim.corpora.dictionary import Dictionary
from numpy import array

import logging
from tqdm import tqdm
from pprint import pprint

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD
from sklearn.decomposition import NMF
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer 

df = pd.read_pickle('pizza_sentences_sample.pkl')

bus_df = pd.read_pickle('business.pkl')


df.head(1)

**Sub-topics within each main topic**

**Groupby topic/subtopic with average sentiment value and frequency of topic/subtopic**

name = "Pieology Pizzeria"
business_id = df['business_id'][df['name']==name][0]
stars = bus_df['stars'][bus_df['business_id'] == 'business_id']
business_id

stars = pd.Series(bus_df['stars'][bus_df['business_id'] == business_id]).iloc[0]
stars

counts = df.groupby(['business_id']).size().reset_index(name='counts')
aggs = pd.merge (df, counts , how = 'left', on='business_id')


aggs = aggs.groupby(['business_id','topic','topic_name']).agg({'sentiment_val':'mean', 
                                                'index':lambda x: x.nunique(),
                                                'counts': 'mean'                                                
                                               }).reset_index()
aggs.columns = ['business_id','topic','topic_name','avg_sentiment', 'topic_count', 'sentence_count']
aggs['importance'] = aggs['topic_count']/aggs['sentence_count']
aggs = aggs.sort_values(by = ['business_id', 'importance'], ascending = False)
aggs.head()

list(aggs['topic'][aggs['business_id'] == 'zmfa8RNxqSuOJ9Nx-CMs_Q'])

**Topic order of importance for each restaurant**  
*For each restaurant, get list of topic/subtopics by order of frequency*  
*Return infograph of average sentiment for each topic/subtopic*
  
**Representative sentence in each topic/subtopic**  
*For each restaurant topic/subtopic item, get representative sentence*    
*Append sentences to summary paragraph in order of importance*

nlp = spacy.load('en', disable=['parser', 'ner'])
def clean_sentences(sentences):
    sentences = [re.sub("\n|\r", "", s) for s in sentences]    
    sentences = [s.lower() for s in sentences] # lower case    
    sentences = [s.translate(str.maketrans('', '', string.punctuation)) for s in sentences] # remove punctuation
    def lemmatize_alphanumeric(sentence):
        words = [lemmatizer.lemmatize(word) for word in sentence.split() if word.isalpha()
                and nlp.vocab[word].is_stop == False]
        sentence = ' '.join(words)
        return(sentence)    
    sentences = [lemmatize_alphanumeric(s) for s in sentences] # lemmatize and remove non alpha-numeric characters
    return(sentences)

sentences = ['the crust is great', 'great authentic pizza', 'they have the best crust',
             'the sauce is great and the crust is so doughy']
mat = get_matrix(sentences)

for i in mat:
    print(i, i.mean())

def get_matrix(sentences):
    from nltk.corpus import stopwords
    stopwords = set(stopwords.words("english"))
    cv = CountVectorizer(stop_words=stopwords)
    
    X_cv = cv.fit_transform(sentences)    
    countvector = pd.DataFrame(X_cv.toarray(), columns=cv.get_feature_names())    
    cosines = cosine_similarity(X_cv)
    
    return(cosines)

def get_top_sentences(cosines, n):
    avg_cosines = []
    for c in range(len(cosines)):
        avg_cosines.append(cosines[c].mean())
    
    top_sentence_dict = {}
    for i in sorted(range(len(avg_cosines)), key=lambda i: avg_cosines[i])[-n:]:
        top_sentence_dict[i] = avg_cosines[i]
    
    top_sentence_i = []
    for w in sorted(top_sentence_dict, key=top_sentence_dict.get, reverse=True):
        top_sentence_i.append(w)
    
    return(top_sentence_i)

business_id = 'zmfa8RNxqSuOJ9Nx-CMs_Q'
topic = 2
n = 2

sentences = list(df['sentence'][(df['business_id'] == business_id) & (df['topic'] == topic)])
text = clean_sentences(sentences)
cosines = get_matrix(text)

top_sentence_i = get_top_sentences(cosines, n)

rep_sentences = []
for i in top_sentence_i:
    rep_sentences.append(sentences[i].lstrip())
    
rep_sentences


def get_rep_sentences(business_id, topic, n):

    sentences = list(df['sentence'][(df['business_id'] == business_id) & (df['topic'] == topic)])
    
    text = clean_sentences(sentences)
    
    cosines = get_matrix(text)
    
    top_sentence_i = get_top_sentences(cosines, n)
    
    rep_sentences = []
    for i in top_sentence_i:
        rep_sentences.append(sentences[i].lstrip())
    
    return(rep_sentences)


def get_rep_sentences_tname(business_id, topic_name, n):
    sentences = list(df['sentence'][(df['business_id'] == business_id) & (df['topic_name'] == topic_name)])    
    text = clean_sentences(sentences)    
    cosines = get_matrix(text)
    top_sentence_i = get_top_sentences(cosines, n)    
    rep_sentences = []
    for i in top_sentence_i:
        rep_sentences.append(sentences[i].lstrip())    
    return(rep_sentences)

def print_rep_sentences(business_id, topic, n):
    rep_sentences = '. '.join(get_rep_sentences(business_id, topic, n))
    return(rep_sentences)

df.topic.unique()

get_rep_sentences('zmfa8RNxqSuOJ9Nx-CMs_Q', 2, 5)

print_rep_sentences('zmfa8RNxqSuOJ9Nx-CMs_Q', 2, 5)

get_rep_sentences_tname('zmfa8RNxqSuOJ9Nx-CMs_Q', 'pizza quality and style', 5)

def get_aggs(business_id, df):
    
    bus = df[df['business_id'] == business_id]
    
    sentence_count = bus.groupby(['business_id']).size().reset_index(name='counts')
    
    bus = pd.merge (bus, sentence_count , how = 'left', on='business_id')
    aggs = bus.groupby(['business_id','topic','topic_name']).agg({'sentiment_val':'mean', 
                                                 'index':lambda x: x.nunique(),
                                                 'counts': 'mean'                                                
                                                }).reset_index()
    aggs.columns = ['business_id','topic','topic_name','avg_sentiment', 'topic_count', 'sentence_count']
    aggs['importance'] = aggs['topic_count']/aggs['sentence_count']
    aggs = aggs.sort_values(by = ['business_id', 'importance'], ascending = False)
    
    return(aggs)

def get_aspect_aggs(name, df):
    business_id = df['business_id'][df['name']==name][0]
    aggs = get_aggs(business_id, df)
    aggs['sentiment_score'] = [(100*X/3.0) for X in aggs['avg_sentiment']]
    aggs = pd.Series(aggs.sentiment_score.values,index=aggs.topic_name).to_dict()
    return(aggs)

def normalize(df, cols):    
    for col in cols:
        max_value = df[col].max()
        min_value = df[col].min()
        df[col] = [(X - min_value) / (max_value - min_value) for X in df[col]]
    return (df)
df = normalize(df, ['sentiment_val'])
df.describe()

aggs = get_aspect_aggs("Pieology Pizzeria", df)
aggs

business_id = 'zmfa8RNxqSuOJ9Nx-CMs_Q'
results_all = []
for k, v in aggs.items():
    results_dict = {}
    results_dict['aspect'] = k
    results_dict['score'] = v
    results_dict['sentences'] = get_rep_sentences_tname(business_id, k, 3)
    results_all.append(results_dict)
results_all

list(df['business_id'][df['name']=='La Famiglia'])[0]

def get_summary(name, df):
    business_id = list(df['business_id'][df['name']==name])[0]
    
    aggs = get_aggs(business_id, df)
    
    summary = []
    for t in list(aggs['topic'][aggs['business_id'] == business_id]):
        summary.append(get_rep_sentences(business_id, t, 1)[0])
    summary = '. '.join(summary)
    return(summary)

get_summary("La Famiglia", df)

**Aspect overview**

df['sentiment_val'].describe()

# def get_topics_lda(tokenized_text, num_topics):
#     id2word = gensim.corpora.Dictionary(tokenized_text)
#     corpus = [id2word.doc2bow(t) for t in tokenized_text]
#     lda_model = gensim.models.ldamulticore.LdaMulticore(corpus=corpus,
#                                            id2word=id2word,
#                                            num_topics=num_topics, 
#                                            random_state=100,
#                                            chunksize=100,
#                                            passes=128,
#                                            per_word_topics=True)
#     return(lda_model)

# df_subs = pd.DataFrame()
# for t in df['topic']:
#     print('topic:' t)
    
#     df_tmp = df['cleaned'][df['topic'] == t]
    
#     lda_model = get_topics_lda(df_tmp, 3)
    
#     lda_model.print_topics(num_words=20)
    
#     pd.to_pickle(lda_model, 'lda_subtopics_20k_'+str(t)+'.pkl')
    
#     get_document_topics = [lda_model.get_document_topics(item) for item in corpus]
#     topics_dict = {}
#     for i, doc in enumerate(get_document_topics):
#         topics_dict[i] = max(doc,key=itemgetter(1))[0]
#     main_topic = pd.DataFrame(list(topics_dict.items()), columns = ['index','sub_topic'+str(t)])
    
#     df_tmp = pd.concat([df_tmp.reset_index(), main_topic['sub_topic'+str(t)]], axis=1)
    
#     df_subs = pd.concat[df_subs, df_tmp]
    
#     del(df_tmp)

#     print('completed')
